Set custom subject headers for DEV mode
