+++ emails/template.py
@@ -1,20 +1,7 @@
 """
 Email template base model
 """
-import os
 from services.email_service import EmailService
-
-
-class EmailAddress:
-    """
-    Common email address
-    """
-
-    enable_to_production = os.getenv("PRODUCTION")
-    cms_talk = "cmstalk+relval@dovecotmta.cern.ch"
-    dev_forum = "cmstalk+test@dovecotmta.cern.ch"
-    # dev_forum = "pdmvserv@cern.ch"
-    forum = cms_talk if enable_to_production else dev_forum
 
 
 def format_new_line(html):
