+++ utils/email_content.py
@@ -3,10 +3,23 @@ Some auxiliary functions
 to parse markdown content before
 transform it into HTML
 """
+import os
 import re
 
 MARKDOWN_ATTACHMENT = r"!\[(.*?)\]\((.*?)\)"
 markdown_parser = re.compile(MARKDOWN_ATTACHMENT)
+
+
+class EmailAddress:
+    """
+    Common email address
+    """
+
+    enable_to_production = os.getenv("PRODUCTION")
+    cms_talk = "cmstalk+relval@dovecotmta.cern.ch"
+    # dev_forum = "cmstalk+test@dovecotmta.cern.ch"
+    dev_forum = "pdmvserv@cern.ch"
+    forum = cms_talk if enable_to_production else dev_forum
 
 
 def parse_attachment_links(content: str, reference_placeholder=True):
